extract afterGenesisStateDigestHex to monetary config
