+++ src/main/resources/application.conf
@@ -54,6 +54,8 @@ ergo {
       epochLength = 2160
       # number of coins reward decrease every epochs (3 Ergo)
       oneEpochReduction = 300000000
+      # Base16 representation of state roothash after genesis
+      afterGenesisStateDigestHex = "b9deff2a27d270e0978e1814f9a62d1bf8092521042fd2b0c531b59629c586ff01"
     }
 
     # Desired time interval between blocks
