+++ src/test/scala/org/ergoplatform/ErgoSanityDigest.scala
@@ -10,7 +10,7 @@ class ErgoSanityDigest extends ErgoSanity[DIGEST_ST] {
 
   override val stateGen: Gen[WrappedDigestState] = {
     boxesHolderGen.map(WrappedUtxoState(_, createTempDir, emission, None)).map { wus =>
-      val digestState = DigestState.create(Some(wus.version), Some(wus.rootHash), createTempDir, settings.nodeSettings)
+      val digestState = DigestState.create(Some(wus.version), Some(wus.rootHash), createTempDir, settings)
       new WrappedDigestState(digestState, wus, settings)
     }
   }
