+++ src/test/scala/org/ergoplatform/utils/ErgoTestHelpers.scala
@@ -14,20 +14,9 @@ import scorex.testkit.utils.FileUtils
 
 import scala.concurrent.ExecutionContext
 
-trait ErgoTestHelpers extends TestkitHelpers with FileUtils with Matchers with ChainGenerator with ErgoGenerators {
+trait ErgoTestHelpers extends ValidBlocksGenerators {
 
   val timeProvider: NetworkTimeProvider = ErgoTestHelpers.defaultTimeProvider
-
-  def createUtxoState(nodeViewHolderRef: Option[ActorRef] = None): (UtxoState, BoxHolder) = {
-    ErgoState.generateGenesisUtxoState(createTempDir, emission.settings, nodeViewHolderRef)
-  }
-
-
-  def createUtxoState(bh: BoxHolder): UtxoState =
-    UtxoState.fromBoxHolder(bh, createTempDir, emission, None)
-
-  def createDigestState(version: VersionTag, digest: ADDigest): DigestState =
-    DigestState.create(Some(version), Some(digest), createTempDir, ErgoSettings.read(None).nodeSettings)
 
 }
