+++ src/test/scala/org/ergoplatform/utils/HistorySpecification.scala
@@ -1,10 +1,10 @@
-package org.ergoplatform.nodeView.history
+package org.ergoplatform.utils
 
 import org.ergoplatform.mining.DefaultFakePowScheme
+import org.ergoplatform.nodeView.history.ErgoHistory
 import org.ergoplatform.nodeView.history.storage.modifierprocessors.blocktransactions.EmptyBlockTransactionsProcessor
 import org.ergoplatform.nodeView.state.StateType
 import org.ergoplatform.settings._
-import org.ergoplatform.utils.ErgoPropertyTest
 import org.scalacheck.Gen
 import scorex.core.settings.ScorexSettings
 
@@ -46,7 +46,8 @@ trait HistorySpecification extends ErgoPropertyTest {
       PoPoWBootstrap, minimalSuffix, mining = false, miningDelay, offlineGeneration = false)
     val scorexSettings: ScorexSettings = null
     val testingSettings: TestingSettings = null
-    val chainSettings = ChainSettings(blockInterval, epochLength, useLastEpochs, DefaultFakePowScheme, MonetarySettings())
+    val chainSettings = ChainSettings(blockInterval, epochLength, useLastEpochs, DefaultFakePowScheme,
+      settings.chainSettings.monetary)
 
     val dir = createTempDir
     val fullHistorySettings: ErgoSettings = ErgoSettings(dir.getAbsolutePath, chainSettings, testingSettings,
